[move-viz] add bi-directional module bundle
