+++ language/tools/move-visibility/src/lib.rs
@@ -1,5 +1,5 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-mod pkg;
+pub mod pkg;
 pub mod pkg_stdlib;
